refactor: gossip poi refactoring and add benches
